+++ CryptoPanel/OptionEditBox.py
@@ -1,0 +1,2 @@
+from PyQt5 import QtWidgets, QtCore, QtGui
+from ui_Widgets import uni_Widget
