--- ui_Widgets/nodeeditor/__init__.py
@@ -1,4 +1,0 @@
-# -*- coding: utf-8 -*-
-
-__author__ = 'Reverier'
-__version__ = '1.2.0'
